cv_interact {ClassHighOrderInteract}	R Documentation

Cross-validation with 'comp_train_pred_interact'

Description

This function does cross-validation prediction on a training data using `comp_train_pred_interact'

Usage

cv_interact(
        ###################### Specify data,order,no_fold ################
        train_x,train_y,no_cls=c(),nos_fth=c(),no_fold=10,order,
        ptn_file,mc_file,do_comp=1,
        ###################### specify for priors  #######################
        alpha=1,sigma_precisions=c(),sigma_means=c(),
        ################# specify for mc sampling ########################
        iters_mc=200,iters_bt=10, iters_sgm=10,w_bt=50,w_sgm=50,m_bt=50,
        m_sgm=50,ini_sigmas=c(),
        ################### specify for prediction #######################
        pred_file = c(),iter_b = 100,forward = 1,iters_pred = 100)

Arguments

train_x 	 Categorical inputs (also called features,covariates,independent variables,explanary variables,predictors) of test data on which your prediction is based. The row is subject and the column is input. Must be coded with 1,2,... (0 is reserved for other use).

train_y 	 Response of training data, a vector of length equal to the row of `train_x'. Must be coded with 1,2,... `no_cls' .

no_cls 	 a number indicating the number of possibilities (classes) of response, by default the maximum value in `train_y'.

nos_fth 	 a vector indicating the number of possibilites (classes) of each feature, by default the maximum value of each feature.

order 	 the order of interactions one want to consider, by default the total number of features, i.e. ncol(train_x).

no_fold 	 Number of folds in Cross-Validation

ptn_file 	 a character sequence specifying the name of the file to which the compression result is sent. The information of the resulting file can be read using function `display_compress_interact'.

new_compression 	 `new_compression=TRUE' indicates doing the compression again.

do_comp 	 `do_comp=1' indicates doing compression, otherwise using original parameterization. This is used only to test program.

alpha 	 `alpha=1' means that Cauchy prior is used, `alpha=2' means that Gaussian prior is used.

sigmas_precisions 	 A vector of precision parameters of the Inverse-Chisquare prior for `sigmas', starting with order 0 (intercept), having length of `order'+1.By default, 1e20 for intercept and 1.5 for others.

sigmas_means 	 A vector of mean parameters of the Inverse-Chisquare prior for `sigmas', of the same format as `sigmas_precisions'. By default, if `Cauchy' prior is used, 0.5 for intercept and 0.005/order for others, if `Gaussian' prior is used, 100 for intercept, 0.05/order for others

mc_file 	 A character sequence specifying the name of the file to which Markov chain is sent

start_over 	 `start_over=TRUE' means that the old Markov chain file is deleted before a Markov chain sampling starts, otherwise the Markov chain will continue with the last iteration stored in it.

iters_mc 	 Number of iterations of Markov chain that will be run. It could be 0, meaning no Markov chain sampling will be run.

iters_bt 	 Number of iterations of slice sampling for `betas' for each iteration of Markov chain sampling

iters_sgm 	 Number of iterations of slice sampling for `sigmas' for each iteration of slice sampling for `betas'

w_bt 	 Width parameter for slice sampling for `betas', i.e. the length of itervals of every stepping out.

w_sgm 	 Width parameter for slice sampling for `sigmas'.

m_bt 	 Maximum number of stepping out for slice sampling for `betas'.

m_sgm 	 Maximum number of stepping out for slice sampling for `sigmas'.

ini_sigmas 	 Initial values for `sigmas', by default `sigmas_means'.

pred_file 	 A character sequence specifying the file to which the prediction result is sent. If `pred_file=c()', the prediction result is printed out on screen (sent to standard output).

iter_b 	 The first iteration of Markov chain used to make prediction.

forward 	 1 of `forward' Markov chain starting from `iter_b' is used to make prediction.

iter_n 	 The last iteration of Markov chain used to make prediction.

Value

A data frame with first 'no_cls' columns being the predictive probability and the next column being the predicted response value is returned.

Author(s)

Longhai Li, Department of Statistics, University of Toronto, longhai@stat.utoronto.ca, http://www.utstat.toronto.edu/longhai

See Also

comp_train_pred_interact, display_compress_interact, read_mc_g, gendata_interact, evaluate_prediction

[Package ClassHighOrderInteract version 1.0 Index]