display_compress_interact {ClassHighOrderInteract}	R Documentation

Display Compression Information of 'comp_train_pred_interact'

Description

This function displays the compression information saved by `comp_train_pred_interact'

Usage

display_compress_interact(file, gids)

Arguments

file 	 A character sequence specifying the name of the file containing the compression information.

gids 	 A vector giving the ID of the groups whose information you want to disaplay. It can be NULL, meaning you want to display only the general information.

Value

It returns a vector of 4 elements: the 1st one is the number of training cases, the 2nd one is `order' considered, which is also length of predictor sequence, the 3rd one is the number of groups of parameters after compression, the 4th one is the number of original patterns expressed in training data

Author(s)

Longhai Li, Department of Statistics, University of Toronto, longhai@stat.utoronto.ca, http://www.utstat.toronto.edu/longhai

See Also

comp_train_pred_interact

Examples

#display only general inforamtion
## Not run:  display_compress_interact("ptn.log",c()) 
#dispaly information for group 0,1,2
## Not run:  display_compress_interact("ptn.log",c(0,1,2))

[Package ClassHighOrderInteract version 1.0 Index]