read_betas_g {ClassHighOrderInteract}	R Documentation

Read 'beta' values from Markov chain log file saved by 'comp_train_pred_interact'

Description

This function reads `beta' values saved by `comp_train_pred_interact'

Usage

read_betas_g(mc_file,ix_g,ix_cls,no_cls,iter_b,forward,n,quiet=1)

Arguments

mc_file 	 Name of the file with Markov chain iterations

ix_g 	 Index of group

ix_cls 	 Index of Class

no_cls 	 Total number of Classes for response

iter_b 	 The first iteration one like to read

forward 	 The number of iteractions skipped

n 	 The number of iterations one like to read

quiet 	 If FALSE print out some relevant information

Value

It returns the values of `beta' one requests with a vector of length `n'.

Author(s)

Longhai Li, Department of Statistics, University of Toronto, longhai@stat.utoronto.ca, http://www.utstat.toronto.edu/longhai

See Also

comp_train_pred_interact,read_mc_g

[Package ClassHighOrderInteract version 1.0 Index]