read_mc_g {ClassHighOrderInteract}	R Documentation

Read Markov chain values from log file saved by 'comp_train_pred_interact'

Description

This function reads Markov chain values saved by `comp_train_pred_interact'

Usage

read_mc_g(group=c("lprobs","sigmas","betas"), 
          ix, mc_file,iter_b,forward,n,quiet)

Arguments

group 	 A character sequence giving the group name of values. It can be one of ``lprobs'',``sigmas'',``betas''.
Group ``lprobs'' includes, in order, log likelihood (ix=0), log prior distribution of ``betas'' (ix=1), log prior distribution of ``sigmas'' (ix=2), log posterior distribution (i.e. sum of former three term)(ix=3).
Group ``sigmas'' includes the values of hyperparameters associated with each order, starting from order 0 (ix=0).
Group ``betas'' includes the values of ``betas''. To read value of betas, please use function `read_betas_g', which is a wrapper of this function for only ``betas''.

ix 	 index of parameters inside group. Starting from 0.

mc_file 	 Name of the file with Markov chain iterations

iter_b 	 The first iteration one like to read

forward 	 The number of iteractions skipped

n 	 The number of iterations one like to read

quiet 	 If FALSE print out some relevant information

Value

It returns the values of parameters one requests with a vector of length `n'.

Author(s)

Longhai Li, Department of Statistics, University of Toronto, longhai@stat.utoronto.ca, http://www.utstat.toronto.edu/longhai

See Also

comp_train_pred_interact,read_betas_g

Examples

## Not run:  read_mc_g("lprobs",0,"mc.log",10,5,400,1) 

[Package ClassHighOrderInteract version 1.0 Index]